Started work on websocket connections and messaging
Will lilkely need to add AsyncAPI documentation later to document messages for each connection.

This probably will not be a replacement for the OpenAPI document, but instead supplement it.

Future work is also needed to make this information available using HTTP as well, such as to allow adding a reset button.